chore(flags): rm feature flag cta from FE (#101015)
rm the feature flag CTA (was only ever shown to LD users)
