+++ static/app/components/events/featureFlags/testUtils.tsx
@@ -154,7 +154,7 @@ export const EMPTY_STATE_SECTION_PROPS = {
   group: GroupFixture(),
 };
 
-export const NO_FLAG_CONTEXT_SECTION_PROPS_NO_CTA = {
+export const NO_FLAG_CONTEXT_SECTION_PROPS = {
   event: EventFixture({
     id: 'abc123def456ghi789jkl',
     contexts: {other: {}},
@@ -164,17 +164,7 @@ export const NO_FLAG_CONTEXT_SECTION_PROPS_NO_CTA = {
   group: GroupFixture({platform: 'unity'}),
 };
 
-export const NO_FLAG_CONTEXT_SECTION_PROPS_CTA = {
-  event: EventFixture({
-    id: 'abc123def456ghi789jkl',
-    contexts: {other: {}},
-    platform: 'javascript',
-  }),
-  project: ProjectFixture({platform: 'javascript', hasFlags: false}),
-  group: GroupFixture({platform: 'javascript'}),
-};
-
-export const NO_FLAG_CONTEXT_WITH_FLAGS_SECTION_PROPS_NO_CTA = {
+export const NO_FLAG_CONTEXT_WITH_FLAGS_SECTION_PROPS = {
   event: EventFixture({
     id: 'abc123def456ghi789jkl',
     contexts: {other: {}},
