+++ static/app/views/issueDetails/groupFeatureFlags/flagDrawerCTA.tsx
@@ -1,7 +1,6 @@
-import {
+import FeatureFlagCTAContent, {
   BannerWrapper,
-  FeatureFlagCTAContent,
-} from 'sentry/components/events/featureFlags/cta/featureFlagInlineCTA';
+} from 'sentry/components/events/featureFlags/cta/featureFlagCTAContent';
 import {useFeatureFlagOnboarding} from 'sentry/components/events/featureFlags/onboarding/useFeatureFlagOnboarding';
 import {useDrawerContentContext} from 'sentry/components/globalDrawer/components';
 import type {PlatformKey} from 'sentry/types/project';
